+++ src/main/java/learn/tinyioc/factory/BeanFactory.java
@@ -9,5 +9,6 @@ import learn.tinyioc.BeanDefinition;
  */
 public interface BeanFactory {
     Object getBean(String name);
-    void registerBeanDefinition(String name, BeanDefinition beanDefinition);
+
+    void registerBeanDefinition(String name, BeanDefinition beanDefinition) throws Exception;
 }