+++ src/test/java/com/learn/ioc/BeanFactoryTest.java
@@ -1,7 +1,9 @@
 package com.learn.ioc;
 
-import com.learn.ioc.service.HelloServiceImpl;
+import com.learn.ioc.service.HelloWorldService;
 import learn.tinyioc.BeanDefinition;
+import learn.tinyioc.PropertyValue;
+import learn.tinyioc.PropertyValues;
 import learn.tinyioc.factory.AutowireCapableBeanFactory;
 import learn.tinyioc.factory.BeanFactory;
 
@@ -12,18 +14,24 @@ import learn.tinyioc.factory.BeanFactory;
  */
 public class BeanFactoryTest {
 
-    public static void main(String[] args) {
+    public static void main(String[] args) throws Exception {
         // 1.初始化beanfactory(系统行为)
         BeanFactory beanFactory = new AutowireCapableBeanFactory();
 
         // 2.注入bean(系统行为)
         BeanDefinition beanDefinition = new BeanDefinition();
-        beanDefinition.setBeanClassName("com.learn.ioc.service.HelloServiceImpl");
-        beanFactory.registerBeanDefinition("helloServiceImpl", beanDefinition);
+        beanDefinition.setBeanClassName("com.learn.ioc.service.HelloWorldService");
+        // 3.设置属性
+        PropertyValues propertyValues = new PropertyValues();
+        propertyValues.addPropertyValue(new PropertyValue("text", "Hello World!"));
+        beanDefinition.setPropertyValues(propertyValues);
 
-        // 3.获取bean(用户行为)
-        HelloServiceImpl helloService = (HelloServiceImpl) beanFactory.getBean("helloServiceImpl");
-        helloService.hello();
+        // 4.生成bean
+        beanFactory.registerBeanDefinition("helloWorldService", beanDefinition);
+
+        // 5.获取bean
+        HelloWorldService helloWorldService = (HelloWorldService) beanFactory.getBean("helloWorldService");
+        helloWorldService.helloWorld();
 
     }
 }